+++ pulsar-client-cpp/python/pulsar_test.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 #
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
@@ -1178,6 +1178,16 @@ class PulsarTest(TestCase):
         except pulsar.PulsarException:
             # Expected
             pass
+
+    def test_listener_name_client(self):
+        client = Client(self.serviceUrl, listener_name='test')
+        try:
+            producer = client.create_producer("persistent://public/default/partitioned_topic_name_test")
+            self.fail()
+        except pulsar.PulsarException:
+            # Expected
+            pass
+        client.close()
 
     def test_negative_acks(self):
         client = Client(self.serviceUrl)