+++ site2/website/release-notes.md
@@ -1026,7 +1026,7 @@ It also includes fixes for some breaking changes introduced in 2.9.0.
 - Cancel scheduled tasks when deleting ManagedLedgerImpl [#12565](https://github.com/apache/pulsar/pull/12565)
 - Add git branch information for PulsarVersion [#12541](https://github.com/apache/pulsar/pull/12541)
 - Websocket should pass the encryption context to consumers [#12539](https://github.com/apache/pulsar/pull/12539)
-- The count of topics on the bundle is less than 2，skip split [#12527](https://github.com/apache/pulsar/pull/12527)
+- The count of topics on the bundle is less than 2, skip split [#12527](https://github.com/apache/pulsar/pull/12527)
 - Fix the reader skips compacted data which original ledger been removed [#12522](https://github.com/apache/pulsar/pull/12522)
 - Fix `messageDedup` delete inactive producer name [#12493](https://github.com/apache/pulsar/pull/12493)
 - Optimize the code: remove extra spaces [#12470](https://github.com/apache/pulsar/pull/12470)